fix: add missing import reference for _stacks-mixins
This adds a import reference to @stackoverflow/stacks/lib/css/exports/_stacks-mixins that is
missing in the production custom-components.less file that causes `npm run build` to fail. This
import is in the development site's site.less file which is why `npm run start` and
`npm run build:dev` succeeded.