ci(beta): consolidate release workflows in a single one
The changeset/action uses internally github.context.sha to reset to the base commit it creates the PR from and that value is always the last commit of the default branch for workflow triggered by workflow_run. This means that the beta release PR that were created were based off main instead of the latest commit of the beta branch.

https://github.com/changesets/action/blob/main/src/git.ts#L100
https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#workflow_run